+++ packages/assignment-10/src/basic/components/SearchDialog/index.tsx
@@ -1,7 +1,6 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import {
   Box,
-  Button,
   HStack,
   Modal,
   ModalBody,
@@ -10,8 +9,6 @@ import {
   ModalHeader,
   ModalOverlay,
   Table,
-  Tbody,
-  Td,
   Text,
   Th,
   Thead,
@@ -19,14 +16,14 @@ import {
   VStack,
 } from '@chakra-ui/react';
 import { useScheduleContext } from '../../ScheduleContext.tsx';
-import { Lecture } from '../../types.ts';
+import { Lecture, SearchOption } from '../../types.ts';
 import { parseSchedule } from "../../utils.ts";
-import axios from "axios";
 import TimeView from './TimeView/index.tsx';
 import MajorView from './MajorView/index.tsx';
 import Forms from './Forms/index.tsx';
-
-const PAGE_SIZE = 100;
+import LectureView from './LectureView/index.tsx';
+import { fetchAllLectures } from '../../utils/fetchAllLectures.ts';
+import { PAGE_SIZE } from '../../constants.ts';
 
 interface Props {
   searchInfo: {
@@ -36,53 +33,6 @@ interface Props {
   } | null;
   onClose: () => void;
 }
-
-interface SearchOption {
-  query?: string,
-  grades: number[],
-  days: string[],
-  times: number[],
-  majors: string[],
-  credits?: number,
-}
-
-const fetchMajors = () => axios.get<Lecture[]>('/schedules-majors.json');
-const fetchLiberalArts = () => axios.get<Lecture[]>('/schedules-liberal-arts.json');
-
-// TODO: 이 코드를 개선해서 API 호출을 최소화 해보세요 + Promise.all이 현재 잘못 사용되고 있습니다. 같이 개선해주세요.
-const fetchAllLectures = async () => {
-  // 클로저를 이용한 API 호출 여부 캐시
-  let hasFetchedMajors = false;
-  let hasFetchedLiberalArts = false;
-
-  const cachedFetchMajors = async() => {
-    if(hasFetchedMajors){
-      return
-    }
-    hasFetchedMajors = true;
-    return fetchMajors()
-  }
-
-  const cachedFetchLiberalArts = async() => {
-    if(hasFetchedLiberalArts){
-      return
-    }
-    hasFetchedLiberalArts = true;
-    return fetchLiberalArts()
-  }
-
-  const response = await Promise.all([
-    (console.log('API Call 1', performance.now()), cachedFetchMajors()),
-    (console.log('API Call 2', performance.now()), cachedFetchLiberalArts()),
-    (console.log('API Call 3', performance.now()), cachedFetchMajors()),
-    (console.log('API Call 4', performance.now()), cachedFetchLiberalArts()),
-    (console.log('API Call 5', performance.now()), cachedFetchMajors()),
-    (console.log('API Call 6', performance.now()), cachedFetchLiberalArts()),
-  ])
-
-  return response.filter(item => item !== undefined)
-};
-
 
 // TODO: 이 컴포넌트에서 불필요한 연산이 발생하지 않도록 다양한 방식으로 시도해주세요.
 const SearchDialog = ({ searchInfo, onClose }: Props) => {
@@ -101,7 +51,6 @@ const SearchDialog = ({ searchInfo, onClose }: Props) => {
     majors: [],
   });
 
-  const allMajors = useMemo(() => [...new Set(lectures.map(lecture => lecture.major))], [lectures]);
   // 메모이제이션을 통해 원하는 시점에만 값의 변경을 발생시킨다.
   const filteredLectures = useMemo(() => {
     const { query = '', credits, grades, days, times, majors } = searchOptions;
@@ -131,6 +80,7 @@ const SearchDialog = ({ searchInfo, onClose }: Props) => {
 
   const lastPage = useMemo(() => Math.ceil(filteredLectures.length / PAGE_SIZE), [filteredLectures]);
   const visibleLectures = useMemo(() => filteredLectures.slice(0, page * PAGE_SIZE), [filteredLectures, page]);
+  const allMajors = useMemo(() => [...new Set(lectures.map(lecture => lecture.major))], [lectures]);
 
   const changeSearchOption = useCallback((field: keyof SearchOption, value: SearchOption[typeof field]) => {
     setPage(1);
@@ -232,23 +182,7 @@ const SearchDialog = ({ searchInfo, onClose }: Props) => {
               </Table>
 
               <Box overflowY="auto" maxH="500px" ref={loaderWrapperRef}>
-                <Table size="sm" variant="striped">
-                  <Tbody>
-                    {visibleLectures.map((lecture, index) => (
-                      <Tr key={`${lecture.id}-${index}`}>
-                        <Td width="100px">{lecture.id}</Td>
-                        <Td width="50px">{lecture.grade}</Td>
-                        <Td width="200px">{lecture.title}</Td>
-                        <Td width="50px">{lecture.credits}</Td>
-                        <Td width="150px" dangerouslySetInnerHTML={{ __html: lecture.major }}/>
-                        <Td width="150px" dangerouslySetInnerHTML={{ __html: lecture.schedule }}/>
-                        <Td width="80px">
-                          <Button size="sm" colorScheme="green" onClick={() => addSchedule(lecture)}>추가</Button>
-                        </Td>
-                      </Tr>
-                    ))}
-                  </Tbody>
-                </Table>
+                <LectureView visibleLectures={visibleLectures} addSchedule={addSchedule}/>
                 <Box ref={loaderRef} h="20px"/>
               </Box>
             </Box>
